Created the PlaceOrder API(POST API)

+++ src/main/java/com/collegemart/model/ProductInventory.java
@@ -1,4 +1,6 @@
 package com.collegemart.model;
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import lombok.*;
 import javax.persistence.*;
 import java.math.BigInteger;
@@ -16,12 +18,11 @@ public class ProductInventory {
 
     private String name;
 
-    private float price;
+    private Double price;
 
     private Long availableQty;
 
     private String description;
-
     @Enumerated(EnumType.STRING)
     private Category category;
 
@@ -30,4 +31,5 @@ public class ProductInventory {
 //            joinColumns = {@JoinColumn(name = "productId")},
 //            inverseJoinColumns = {@JoinColumn(name = "orderId")})
 //    private List<Orders> orders;
+    // @JsonIdentityInfo(property = "id",generator = ObjectIdGenerators.PropertyGenerator.class)
 }
